requirements.txt
boto==2.48.0
boto3==1.7.55
botocore==1.10.55
certifi==2018.4.16
chardet==3.0.4
contextlib2==0.5.5
dj-database-url==0.5.0
Django==1.11
django-oauth-toolkit==1.0.0
django-storages==1.6.6
djangorestframework==3.8.2
docutils==0.14
futures==3.2.0
gunicorn==19.8.1
idna==2.7
jmespath==0.9.3
MySQL-python==1.2.5
oauthlib==2.1.0
psycopg2==2.7.4
python-dateutil==2.7.3
pytz==2018.4
raven==6.9.0
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
